// go-grpc-angular/src/app/grpc/SittelleConst_pb.d.ts
export enum EnumTaskField { 
  TASK_FIELD_UNDEFINED = 0,
  TASK_FIELD_ITEM_PARENT = 1,
  TASK_FIELD_ITEM_SOURCE = 2,
  TASK_FIELD_ITEM_MON_OBJECT = 3,
  TASK_FIELD_ITEM_CHILD_LIST = 4,
  TASK_FIELD_STATUS_CHAIN_ID = 5,
  TASK_FIELD_STATUS_MESSAGE_ID = 6,
  TASK_FIELD_STATUS_TEXT = 7,
  TASK_FIELD_STATUS_COMMENT = 8,
  TASK_FIELD_CONTENT_CAPTION = 9,
  TASK_FIELD_CONTENT_TEXT = 10,
  TASK_FIELD_CONTENT_PRIORITY = 11,
  TASK_FIELD_CONTENT_ACTIVE_PLANNING_START = 12,
  TASK_FIELD_CONTENT_DATE_PLANNING_START = 13,
  TASK_FIELD_CONTENT_ACTIVE_PLANNING_END = 14,
  TASK_FIELD_CONTENT_DATE_PLANNING_END = 15,
  TASK_FIELD_CONTENT_TIME_ESTIMATE = 16,
  TASK_FIELD_CONTENT_MON_OBJECT_LIST = 17,
  TASK_FIELD_STATE_VALUE = 18,
  TASK_FIELD_COMMENT = 19,
}
// ...
export enum EnumManualParamType { 
  MANUAL_TYPE_UNDEFINED = 0,
  MANUAL_TYPE_COMBOBOX = 1,
  MANUAL_TYPE_DATETIME = 2,
  MANUAL_TYPE_STRING = 3,
  MANUAL_TYPE_LIST = 4,
  MANUAL_TYPE_TIME = 5,
  MANUAL_TYPE_BASE64 = 6,
  MANUAL_TYPE_INTEGER = 7,
  MANUAL_TYPE_FLOAT = 8,
  MANUAL_TYPE_DATE = 9,
  MANUAL_TYPE_PERIOD = 10,
  MANUAL_TYPE_UNSIGHED_INT = 11,
  MANUAL_TYPE_UNSIGHED_FLOAT = 12,
  MANUAL_TYPE_BOOL = 13,
  MANUAL_TYPE_MEMO = 14,
}
// ...
export enum EnumTaskRight { 
  TASK_RIGHT_TYPE_LIST_READ = 0,
  TASK_RIGHT_READ = 1,
  TASK_RIGHT_STATUS_APPLY = 2,
  TASK_RIGHT_ITEM_CHANGE = 3,
  TASK_RIGHT_ROLE_CHANGE = 4,
  TASK_RIGHT_STATE_CHANGE = 5,
  TASK_RIGHT_COMMENT_ADD = 6,
  TASK_RIGHT_MANUAL_CHANGE = 7,
  TASK_RIGHT_STATUS_INPUT_CHANGE = 8,
  TASK_RIGHT_CONTENT_CHANGE = 9,
}